hotfix-refactoring folder structure
