+++ src/components/feature/end_user/content/slider.tsx
